requirements.txt
flask
openai
pytest
python-dotenv
requests

